try gold evaluation

+++ src/train_x_encoder.py
@@ -157,7 +157,7 @@ if __name__ == '__main__':
         print(f'Correct count: {correct_count}/{mention_count} = {correct_count/mention_count*100}%')
         dev_loss = dev_loss / len(dev_loader.dataset)
         dev_preds = torch.from_numpy(dev_preds)
-        dev_labels = torch.from_numpy(dev_labels)
+        dev_labels = torch.from_numpy(dev_labels).int()
         threshold = 0.7
         accuracy = BinaryAccuracy(threshold=threshold)
         precision = BinaryPrecision(threshold=threshold)
